Tweepy
Twitter allows access to parts of our service via APIs to allow people to build software that integrates with Twitter, like a solution that helps a company respond to customer feedback on Twitter. ... Our API platform provides broad access to public Twitter data that users have chosen to share with the world.